Feat : modal 구조 변경에 따른 deleteProfile func 수정
